feat: use derived pivot configuration in visualization warning (#19067)
<!-- Thanks so much for your PR, your contribution is appreciated! ❤️ -->

Closes: 

### Description:
Refactored the VisualizationCard component to use a derived pivot configuration instead of directly accessing the unsaved chart version's pivot config. This change introduces a new `dirtyPivotConfiguration` that is calculated using the `derivePivotConfigurationFromChart` function, which provides a more complete pivot configuration object to the VisualizationWarning component.

The VisualizationWarning component was updated to accept this new configuration object instead of just the pivot dimensions array, making the component more flexible and providing it with more context about the pivot configuration.