fix: disable space access in embed mode (#19087)
<!-- Thanks so much for your PR, your contribution is appreciated! ❤️ -->

Related to: #19083 / PROD-2199 

### Description:
Disable "Create in any space" functionality in embed mode by checking for the presence of an embed token. This prevents the hook from fetching space summaries when in embed mode and ensures the hook returns `false` when an embed token is present.

TLDR; this fired a `/spaces` request that would be 401 and also unneeded, since we don't save from embeds.